feat: Enabled -O3 and thin-lto for psycho-nvse

+++ psycho-nvse/src/entry.rs
@@ -125,8 +125,8 @@ pub unsafe extern "C" fn NVSEPlugin_Query(
 
     log::info!("NVSEPlugin_Query called! NVSEInterface address: {:p}", nvse);
 
-    let nvse_version = ExeVersion::from_u32(nvse.nvseVersion as u32);
-    let runtime_version = ExeVersion::from_u32(nvse.runtimeVersion as u32);
+    let nvse_version = ExeVersion::from_u32(nvse.nvseVersion);
+    let runtime_version = ExeVersion::from_u32(nvse.runtimeVersion);
 
     log::info!("NVSE version: {}", nvse_version);
     log::info!("Runtime version: {}", runtime_version);
